发送前清除串口缓存避免排长队； 修改传给绘图函数的gForce数据类型。

+++ oym8CHWave/gfsdkqml.cpp
@@ -262,7 +262,7 @@ void gfSdkQml::onExtendedDeviceData(SPDEVICE device, DeviceDataType dataType, gf
 
         // qvector, must be RawData;
         if (128 == data->size()) {
-            QVector<double> tempRawData;
+            QVector<uint8_t> tempRawData;
             for (int i = 0; i < data->size(); ++i) {
                 tempRawData.push_back((*data)[i]);
             }
